Add more lifecycle documentation
