+++ src/lib.rs
@@ -144,57 +144,10 @@
 mod error;
 pub mod geom;
 pub mod graphics;
-pub mod lifecycle {
-    use crate::graphics::Graphics;
-    use crate::Result;
-    use blinds::run_gl;
-    pub use blinds::{EventStream, Settings, Window, *};
-    use std::future::Future;
-
-    pub fn run<F, T>(settings: Settings, app: F) -> !
-    where
-        T: 'static + Future<Output = Result<()>>,
-        F: 'static + FnOnce(Window, Graphics, EventStream) -> T,
-    {
-        #[cfg(target_arch = "wasm32")]
-        web_logger::custom_init(web_logger::Config {
-            level: log::Level::Debug,
-        });
-        #[cfg(not(target_arch = "wasm32"))]
-        simple_logger::init_with_level(log::Level::Debug)
-            .expect("A logger was already initialized");
-
-        use crate::geom::{Rectangle, Transform};
-
-        let size = settings.size;
-        let screen_region = Rectangle::new_sized(size);
-        run_gl(settings, move |window, ctx, events| {
-            #[cfg(not(target_arch = "wasm32"))]
-            {
-                if let Err(_) = std::env::set_current_dir("static") {
-                    log::warn!("Warning: no asset directory found. Please place all your assets inside a directory called 'static' so they can be loaded");
-                    log::warn!("Execution continuing, but any asset-not-found errors are likely due to the lack of a 'static' directory.")
-                }
-            }
-
-            let ctx = golem::Context::from_glow(ctx).unwrap();
-            let mut graphics = Graphics::new(ctx).unwrap();
-            graphics.set_projection(Transform::orthographic(screen_region));
-
-            async {
-                match app(window, graphics, events).await {
-                    Ok(()) => log::info!("Exited successfully"),
-                    Err(err) => {
-                        log::error!("Error: {:?}", err);
-                        panic!("{:?}", err);
-                    }
-                }
-            }
-        });
-    }
-}
+pub mod lifecycle;
 #[cfg(feature = "saving")]
 pub mod saving {
+    //! A module to manage cross-platform save data via the [`gestalt`] library
     pub use gestalt::*;
 }
 pub use crate::error::QuicksilverError;
